deps/v8/test/inspector/runtime/console-time-end-stack-expected.txt

Checks that console.timeEnd() always attached exactly one stack frame

Running test: top
Stack depth: 1

Running test: deep
Stack depth: 1
